requirements.txt
celery==3.1.23
ipython[notebook]
pillow
